+++ src/react/components/positioning/Positioning.js
@@ -26,21 +26,21 @@ import {
  * @param {Positioning} popup
  */
 function alignBaseWithTrigger(popupRect, triggerRect, popup) {
-	const { gutter, gap } = popup.props;
+	const { viewportGutter, triggerDropdownGap } = popup.props;
 	const spaceAboveTrigger = calcSpaceAbove(triggerRect);
 	const spaceBelowTrigger = calcSpaceBelow(triggerRect);
 	const spaceLeftOfTrigger = calcSpaceLeft(triggerRect);
 	const spaceRightOfTrigger = calcSpaceRight(triggerRect);
 
-	const canGoBelow = popupRect.height <= spaceBelowTrigger - gutter - gap;
-	const canAlignLeftEdge = popupRect.width <= spaceRightOfTrigger + triggerRect.width - gutter - gap;
+	const canGoBelow = popupRect.height <= spaceBelowTrigger - viewportGutter - triggerDropdownGap;
+	const canAlignLeftEdge = popupRect.width <= spaceRightOfTrigger + triggerRect.width - viewportGutter - triggerDropdownGap;
 
 	const placeBelow = canGoBelow || spaceBelowTrigger >= spaceAboveTrigger;
 	const alignLeftEdge = canAlignLeftEdge || spaceRightOfTrigger >= spaceLeftOfTrigger;
 
 	const popupLeftPos = alignLeftEdge
-		? Math.max(triggerRect.left, gutter)
-		: Math.min(Math.max(triggerRect.right - popupRect.width, gutter), verge.viewportW() - popupRect.width - gutter);
+		? Math.max(triggerRect.left, viewportGutter)
+		: Math.min(Math.max(triggerRect.right - popupRect.width, viewportGutter), verge.viewportW() - popupRect.width - viewportGutter);
 
 	const translateX = isNarrowViewport()
 		? '0px'
@@ -52,8 +52,8 @@ function alignBaseWithTrigger(popupRect, triggerRect, popup) {
 	// Initially the gap offset here was done through css calc properties in the translate function. Unfortunately
 	// this caused issues, as calc is invalid as a parameter of translate within IE11
 	const topValue = placeBelow ?
-		triggerRect.top + scrollTopAmount() + gap
-		: triggerRect.top + scrollTopAmount() - gap;
+		triggerRect.top + scrollTopAmount() + triggerDropdownGap
+		: triggerRect.top + scrollTopAmount() - triggerDropdownGap;
 
 	popup.setState({
 		top: topValue,
@@ -91,7 +91,7 @@ function positionOnShow(popup) {
 	// Safety check due to slim chance of unmount during setTimeout duration
 	if (popup.positionEl && document.body.contains(popup.positionEl)) {
 		popup.positionComponent();
-		if (popup.props.setMaxHeight) {
+		if (popup.props.shouldRestrictMaxHeight) {
 			popup.calculateMaxHeight();
 		}
 		// Tell the render method it's OK to render without "visibility: hidden"
@@ -127,7 +127,7 @@ function getDefaultState() {
  * @returns {Object|null}
  */
 function getPositionCalculationStyles(popup) {
-	if (!popup.props.renderHidden && !popup.state.positioned) {
+	if (popup.props.isVisible && !popup.state.positioned) {
 		return {
 			visibility: 'hidden',
 			position: 'fixed',
@@ -154,7 +154,7 @@ class Positioning extends PureComponent {
 	}
 
 	componentDidMount() {
-		if (!this.props.renderHidden) {
+		if (this.props.isVisible) {
 			attachListeners(this);
 		}
 	}
@@ -165,7 +165,7 @@ class Positioning extends PureComponent {
 		// If the popup is going from hidden to visible but hasn't been positioned yet, the render method will ensure
 		// that everything is rendered with "visibility: hidden".  Wait a bit to make sure all children also render,
 		// then measure things and position the popup correctly on the screen.
-		if (!props.renderHidden && !state.positioned) {
+		if (props.isVisible && !state.positioned) {
 			this._positionTimer = setTimeout(positionOnShow, 50, this);
 		}
 
@@ -173,8 +173,8 @@ class Positioning extends PureComponent {
 			this._visibleTimer = setTimeout(props.onVisible, 50);
 		}
 
-		if (props.renderHidden !== prevProps.renderHidden) {
-			if (props.renderHidden) {
+		if (props.isVisible !== prevProps.isVisible) {
+			if (!props.isVisible) {
 				// If we're hiding the popup, reset the state to defaults so that the next show event will properly
 				// reposition everything.
 				this.setState(getDefaultState());
@@ -188,8 +188,8 @@ class Positioning extends PureComponent {
 			}
 		}
 
-		if (prevProps.setMaxHeight !== props.setMaxHeight) {
-			if (props.setMaxHeight) {
+		if (prevProps.shouldRestrictMaxHeight !== props.shouldRestrictMaxHeight) {
+			if (props.shouldRestrictMaxHeight) {
 				this.calculateMaxHeight();
 			} else {
 				this.setState({
@@ -219,7 +219,7 @@ class Positioning extends PureComponent {
 			const baseRect = popup.positionEl && popup.positionEl.firstChild.getBoundingClientRect();
 
 			if (isBaseRendered(baseRect)) {
-				if (!popup.props.forceDesktop && isNarrowViewport()) {
+				if (!popup.props.isNotResponsive && isNarrowViewport()) {
 					alignBottomLeft(popup);
 				} else {
 					alignBaseWithTrigger(baseRect, triggerDOM.getBoundingClientRect(), popup);
@@ -236,11 +236,11 @@ class Positioning extends PureComponent {
 	 */
 	calculateMaxHeight() {
 		const popup = this;
-		const { gutter, parentRef, gap } = popup.props;
+		const { viewportGutter, parentRef, triggerDropdownGap } = popup.props;
 		const triggerDOM = parentRef.firstChild;
 
 		if (verge.inViewport(triggerDOM)) {
-			if (!popup.props.forceDesktop && isNarrowViewport()) {
+			if (!popup.props.isNotResponsive && isNarrowViewport()) {
 				popup.setState({
 					maxHeight: verge.viewportH(),
 				});
@@ -250,7 +250,7 @@ class Positioning extends PureComponent {
 				const spaceBelowTrigger = calcSpaceBelow(triggerRect);
 
 				popup.setState({
-					maxHeight: Math.max(spaceAboveTrigger, spaceBelowTrigger) - gutter - gap,
+					maxHeight: Math.max(spaceAboveTrigger, spaceBelowTrigger) - viewportGutter - triggerDropdownGap,
 				});
 			}
 		}
@@ -263,11 +263,11 @@ class Positioning extends PureComponent {
 	 */
 	getStyles() {
 		const { maxHeight, transform, top, bottom } = this.state;
-		const { matchTriggerWidth, parentRef, forceDesktop } = this.props;
-		const isMobile = isNarrowViewport() && !forceDesktop;
+		const { isTriggerWidthMatched, parentRef, isNotResponsive } = this.props;
+		const isMobile = isNarrowViewport() && !isNotResponsive;
 		let width = null;
 		let maxWidth = null;
-		if (matchTriggerWidth && !isMobile && parentRef != null && parentRef.firstChild != null) {
+		if (isTriggerWidthMatched && !isMobile && parentRef != null && parentRef.firstChild != null) {
 			width = parentRef.firstChild.getBoundingClientRect().width;
 			maxWidth = 'none';
 		}
@@ -284,16 +284,16 @@ class Positioning extends PureComponent {
 
 	render() {
 		const popup = this;
-		const { children, renderHidden } = popup.props;
+		const { children, isVisible } = popup.props;
 		const { positioned } = popup.state;
 		const positioningStyles = getPositionCalculationStyles(popup);
-		const clonedChildren = renderHidden || !positioned ? children : React.cloneElement(children, {
+		const clonedChildren = !isVisible || !positioned ? children : React.cloneElement(children, {
 			className : cn( children.props.className, { 'dropdown-positionabove' : popup.state.alignTop } ),
 			style : popup.getStyles(),
 		});
 
 		return (
-			<Portal isOpened={!renderHidden} onOpen={popup.positionComponent} >
+			<Portal isOpened={isVisible} onOpen={popup.positionComponent} >
 				<div style={positioningStyles} ref={portal => popup.positionEl = portal} className="xui-container">
 					{clonedChildren}
 				</div>
@@ -306,29 +306,29 @@ Positioning.propTypes = {
 	className: PropTypes.string,
 	children: PropTypes.node,
 	/**true when the component is rendered but not displayed */
-	renderHidden: PropTypes.bool,
+	isVisible: PropTypes.bool,
 	/**A DOM object of the parent node. */
 	parentRef: PropTypes.object,
 	/**A buffer value added to measure between the edge of the viewport and the component before flipping its position. */
-	gutter: PropTypes.number,
+	viewportGutter: PropTypes.number,
 	/**A max height will mean an overflowed popup will scroll for the user rather than render outside of the viewport. True by default. */
-	setMaxHeight: PropTypes.bool,
+	shouldRestrictMaxHeight: PropTypes.bool,
 	/** Force the desktop UI, even if the viewport is narrow enough for mobile. */
-	forceDesktop: PropTypes.bool,
+	isNotResponsive: PropTypes.bool,
 	/** The amount of space to put between the trigger and the dropdown */
-	gap: PropTypes.number,
+	triggerDropdownGap: PropTypes.number,
 	/** Callback for when the positioned element becomes visible  */
 	onVisible: PropTypes.func,
 	/** Setting to true will for the dropdown to be as wide as the trigger. */
-	matchTriggerWidth: PropTypes.bool,
+	isTriggerWidthMatched: PropTypes.bool,
 };
 
 Positioning.defaultProps = {
-	gutter: 10,
-	setMaxHeight: true,
-	forceDesktop: false,
-	gap: 5,
-	matchTriggerWidth: false,
+	viewportGutter: 10,
+	shouldRestrictMaxHeight: true,
+	isNotResponsive: false,
+	triggerDropdownGap: 5,
+	isTriggerWidthMatched: false,
 };
 
 export default Positioning;